+++ inst/shiny_app/www/js/iaoreader.js
@@ -3,21 +3,27 @@ let IAOReader = class {
     height = 720;
     margin = 30;
 
-    get svg() {
-        return d3.select("div#plot svg");
+    constructor() {
+        var plot_div = d3.select("div#plot");
+        var svg = plot_div.select("svg");
+
+        // Creating the SVG if it does not exist.
+        if (svg.empty()) {
+            // TODO: remove redundant parameters
+            svg = plot_div.append("svg")
+                .attr("_x", this.width)
+                .attr("_y", this.height)
+                .attr("_margin", this.margin)
+                .attr("viewBox", "0 0 " + this.width + " " + this.height);
+        }
+
+        this.svg = svg;
     }
 
-    // Note: this has to be called manually due to the script being invoked
-    //       before the full HTML creation. Calling update_svg in constructor
-    //       has no effect due to the svg not being created yet.
-    update_svg() {
-        // TODO: remove redundant parameters
-        this.svg
-            .attr("_x", this.width)
-            .attr("_y", this.height)
-            .attr("_margin", this.margin)
-            .attr("viewBox", "0 0 " + this.width + " " + this.height);
-    }
+    // TODO: add plot settings parameters object.
+    // TODO: move axis creation, scales, max_seq_len etc. to this class.
 }
 
-var iaoreader = new IAOReader();
+
+// This variable has valueu assigned by Shiny main server function.
+var iaoreader;