+++ inst/shiny_app/www/js/shiny_handlers/initialize_iaoreader.js
@@ -1,0 +1,3 @@
+Shiny.addCustomMessageHandler("initialize_iaoreader", function(_) {
+    iaoreader = new IAOReader();
+});