+++ inst/shiny_app/www/js/shiny_handlers/plot_settings_title.js
@@ -1,7 +1,6 @@
 Shiny.addCustomMessageHandler("plot_settings_title", function(plot_title_value) {
     var svg = d3.select("div#plot svg");
     if (svg.empty()) return;
-    var margin = svg.attr("_margin");
 
     // Creating the title if it does not exist.
     var plot_title = svg.select("text#plot_title");
@@ -11,7 +10,7 @@ Shiny.addCustomMessageHandler("plot_settings_title", function(plot_title_value) 
                 .attr("id", "plot_title")
                 .attr("text-anchor", "middle")
                 .attr("x", "50%")
-                .attr("y", (margin === null ? 30 : margin)  / 2 + "px");
+                .attr("y", margin / 2 + "px");
     }
 
     plot_title.text(plot_title_value);