+++ inst/shiny_app/www/js/shiny_handlers/update_svg.js
@@ -1,3 +1,0 @@
-Shiny.addCustomMessageHandler("update_svg", function(_) {
-    iaoreader.update_svg();
-});